refactor: cleanup `events/`
