+++ packages/backend/convex/_generated/api.d.ts
@@ -15,13 +15,13 @@ import type * as data_ml_friendRecs from "../data_ml/friendRecs.js";
 import type * as data_ml_friends from "../data_ml/friends.js";
 import type * as data_ml_universal from "../data_ml/universal.js";
 import type * as data_ml_users from "../data_ml/users.js";
-import type * as eventRecs from "../eventRecs.js";
 import type * as eventSeedsStatic from "../eventSeedsStatic.js";
 import type * as events_attendance from "../events/attendance.js";
 import type * as events_ingest from "../events/ingest.js";
 import type * as events_mutations from "../events/mutations.js";
 import type * as events_queries from "../events/queries.js";
 import type * as events_search from "../events/search.js";
+import type * as events_utils from "../events/utils.js";
 import type * as files from "../files.js";
 import type * as friends from "../friends.js";
 import type * as likes from "../likes.js";
@@ -45,13 +45,13 @@ declare const fullApi: ApiFromModules<{
   "data_ml/friends": typeof data_ml_friends;
   "data_ml/universal": typeof data_ml_universal;
   "data_ml/users": typeof data_ml_users;
-  eventRecs: typeof eventRecs;
   eventSeedsStatic: typeof eventSeedsStatic;
   "events/attendance": typeof events_attendance;
   "events/ingest": typeof events_ingest;
   "events/mutations": typeof events_mutations;
   "events/queries": typeof events_queries;
   "events/search": typeof events_search;
+  "events/utils": typeof events_utils;
   files: typeof files;
   friends: typeof friends;
   likes: typeof likes;
